fix overflow

+++ Math.fs
@@ -6,8 +6,8 @@ open System
 /// 32-bit integer, and returns the number of steps it takes to reach 1 in the
 /// Collatz sequence.
 let collatz (n: int) : int =
-    let rec tail (n: int) (steps: int) : int =
+    let rec tail (n: int64) (steps: int) : int =
         if n = 1 then steps
-        elif n % 2 = 0 then tail (n / 2) (steps + 1)
-        else tail (3 * n + 1) (steps + 1)
+        elif n % 2L = 0 then tail (n / 2L) (steps + 1)
+        else tail (3L * n + 1L) (steps + 1)
     tail n 0
